switch some config to context-params, and update with Cru info
This avoids repeated config by use context-params.
The login url isn't repeated, but I like it being next to the casServerUrlPrefix setting, since it's so similar. (It's unfortunate that it can't be derived from casServerUrlPrefix.)

This adds Cru's cas server urls, since this repo is aimed primarily at Cru folks.